renames for conduit lib
