+++ src/libs/conduit/conduit_data_array.cpp
@@ -44,10 +44,10 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: DataArray.cpp
+/// file: conduit_data_array.cpp
 ///
 //-----------------------------------------------------------------------------
-#include "DataArray.hpp"
+#include "conduit_data_array.hpp"
 
 //-----------------------------------------------------------------------------
 // -- standard includes -- 
@@ -58,7 +58,7 @@
 //-----------------------------------------------------------------------------
 // -- conduit includes -- 
 //-----------------------------------------------------------------------------
-#include "Utils.hpp"
+#include "conduit_utils.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
