+++ src/libs/conduit/conduit_data_array.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: DataArray.hpp
+/// file: conduit_data_array.hpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -54,8 +54,8 @@
 //-----------------------------------------------------------------------------
 // -- conduit  includes -- 
 //-----------------------------------------------------------------------------
-#include "Core.hpp"
-#include "DataType.hpp"
+#include "conduit_core.hpp"
+#include "conduit_data_type.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
