+++ src/libs/conduit/conduit_data_type.cpp
@@ -44,16 +44,16 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: DataType.cpp
+/// file: conduit_data_type.cpp
 ///
 //-----------------------------------------------------------------------------
-#include "DataType.hpp"
+#include "conduit_data_type.hpp"
 
 //-----------------------------------------------------------------------------
 // conduit includes
 //-----------------------------------------------------------------------------
-#include "Utils.hpp"
-#include "Schema.hpp"
+#include "conduit_utils.hpp"
+#include "conduit_schema.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
