+++ src/libs/conduit/conduit_data_type.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: DataType.hpp
+/// file: conduit_data_type.hpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -61,8 +61,8 @@
 //-----------------------------------------------------------------------------
 // -- conduit library includes -- 
 //-----------------------------------------------------------------------------
-#include "Core.hpp"
-#include "Endianness.hpp"
+#include "conduit_core.hpp"
+#include "conduit_endianness.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
