+++ src/libs/conduit/conduit_endianness.cpp
@@ -44,10 +44,10 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Endianness.cpp
+/// file: conduit_endianness.cpp
 ///
 //-----------------------------------------------------------------------------
-#include "Endianness.hpp"
+#include "conduit_endianness.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
