+++ src/libs/conduit/conduit_endianness_types.h
@@ -44,111 +44,24 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Core.cpp
+/// file: conduit_endianness_types.h
 ///
 //-----------------------------------------------------------------------------
-#include "Core.hpp"
 
-
-//-----------------------------------------------------------------------------
-// conduit includes
-//-----------------------------------------------------------------------------
-#include "Node.hpp"
-
-// Note: This header is only needed a compile time.
-#include "Conduit_License.hpp"
+#ifndef CONDUIT_ENDIANNESS_TYPES_H
+#define CONDUIT_ENDIANNESS_TYPES_H
 
 //-----------------------------------------------------------------------------
-// -- begin conduit:: --
+/// conduit_endianness_type_id is an Enumeration used to describe the  
+/// endianness cases supported by conduit
 //-----------------------------------------------------------------------------
-namespace conduit
+typedef enum
 {
+    CONDUIT_ENDIANNESS_DEFAULT_ID  = 0, // (machine default)
+    CONDUIT_ENDIANNESS_BIG_ID,          // big endian
+    CONDUIT_ENDIANNESS_LITTLE_ID        // little endian
+} conduit_endianness_type_id;
 
 
-//---------------------------------------------------------------------------//
-std::string
-about()
-{
-    Node n;
-    about(n);
-    return n.to_json();
-}
-
-//---------------------------------------------------------------------------//
-void
-about(Node &n)
-{
-    n.reset();
-    n["version"] = "{alpha}";
-    n["install_prefix"] = CONDUIT_INSTALL_PREFIX;
-    n["license"] = CONDUIT_LICENSE_TEXT;
-    
-    // TODO: include compiler info, etc ?
-
-    // Type Info Map
-    Node &nn = n["native_typemap"];
-
-    // ints
-#ifdef CONDUIT_INT8_NATIVE_NAME
-    nn["int8"] = CONDUIT_INT8_NATIVE_NAME;
-#else
-    nn["int8"] = "<unmapped>";
 #endif
-#ifdef CONDUIT_INT16_NATIVE_NAME
-    nn["int16"] = CONDUIT_INT16_NATIVE_NAME;
-#else
-    nn["int16"] = "<unmapped>";
-#endif
-#ifdef CONDUIT_INT32_NATIVE_NAME
-    nn["int32"] = CONDUIT_INT32_NATIVE_NAME;
-#else
-    nn["int32"] = "<unmapped>";
-#endif
-#ifdef CONDUIT_INT64_NATIVE_NAME
-    nn["int64"] = CONDUIT_INT64_NATIVE_NAME;
-#else
-    nn["int64"] = "<unmapped>";
-#endif
-
-    // unsigned ints
-#ifdef CONDUIT_UINT8_NATIVE_NAME
-    nn["uint8"] = CONDUIT_UINT8_NATIVE_NAME;
-#else
-    nn["uint8"] = "<unmapped>";
-#endif
-#ifdef CONDUIT_UINT16_NATIVE_NAME
-    nn["uint16"] = CONDUIT_UINT16_NATIVE_NAME;
-#else
-    nn["uint16"] = "<unmapped>";
-#endif
-#ifdef CONDUIT_UINT32_NATIVE_NAME
-    nn["uint32"] = CONDUIT_UINT32_NATIVE_NAME;
-#else
-    nn["uint32"] = "<unmapped>";
-#endif
-#ifdef CONDUIT_UINT64_NATIVE_NAME
-    nn["uint64"] = CONDUIT_UINT64_NATIVE_NAME;
-#else
-    nn["uint64"] = "<unmapped>";
-#endif
-
-    // floating points numbers
-#ifdef CONDUIT_FLOAT32_NATIVE_NAME
-    nn["float32"] = CONDUIT_FLOAT32_NATIVE_NAME;
-#else
-    nn["float32"] = "<unmapped>";
-#endif
-#ifdef CONDUIT_FLOAT64_NATIVE_NAME
-    nn["float64"] = CONDUIT_FLOAT64_NATIVE_NAME;
-#else
-    nn["float64"] = "<unmapped>";
-#endif
-
-}
-
-
-}
-//-----------------------------------------------------------------------------
-// -- end conduit:: --
-//-----------------------------------------------------------------------------
