+++ src/libs/conduit/conduit_exports.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Conduit_Exports.hpp
+/// file: conduit_exports.hpp
 ///
 //-----------------------------------------------------------------------------
