+++ src/libs/conduit/conduit_generator.cpp
@@ -44,10 +44,11 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Generator.cpp
+/// file: conduit_generator.cpp
 ///
 //-----------------------------------------------------------------------------
-#include "Generator.hpp"
+#include "conduit_generator.hpp"
+
 
 //-----------------------------------------------------------------------------
 // -- standard lib includes -- 
@@ -63,8 +64,8 @@
 //-----------------------------------------------------------------------------
 // -- conduit library includes -- 
 //-----------------------------------------------------------------------------
-#include "Error.hpp"
-#include "Utils.hpp"
+#include "conduit_error.hpp"
+#include "conduit_utils.hpp"
 
 
 //-----------------------------------------------------------------------------
