+++ src/libs/conduit/conduit_generator.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Generator.hpp
+/// file: conduit_generator.hpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -54,11 +54,9 @@
 //-----------------------------------------------------------------------------
 // -- conduit includes -- 
 //-----------------------------------------------------------------------------
-#include "Core.hpp"
-#include "Endianness.hpp"
-#include "DataType.hpp"
-#include "Node.hpp"
-#include "Schema.hpp"
+#include "conduit_core.hpp"
+#include "conduit_schema.hpp"
+#include "conduit_node.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
