+++ src/libs/conduit/conduit_node.cpp
@@ -44,10 +44,10 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Node.cpp
+/// file: conduit_node.cpp
 ///
 //-----------------------------------------------------------------------------
-#include "Node.hpp"
+#include "conduit_node.hpp"
 
 //-----------------------------------------------------------------------------
 // -- standard cpp lib includes -- 
@@ -79,8 +79,9 @@
 //-----------------------------------------------------------------------------
 // -- conduit includes -- 
 //-----------------------------------------------------------------------------
-#include "Utils.hpp"
-#include "Generator.hpp"
+#include "conduit_error.hpp"
+#include "conduit_utils.hpp"
+
 
 //-----------------------------------------------------------------------------
 //
