+++ src/libs/conduit/conduit_node_iterator.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: NodeIterator.hpp
+/// file: conduit_node_iterator.hpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -54,9 +54,7 @@
 //-----------------------------------------------------------------------------
 // -- conduit includes -- 
 //-----------------------------------------------------------------------------
-#include "Error.hpp"
-#include "Node.hpp"
-#include "Utils.hpp"
+#include "conduit_node.hpp"
 
 //-----------------------------------------------------------------------------
 // -- begin conduit:: --
