+++ src/libs/conduit/conduit_schema.cpp
@@ -44,10 +44,10 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Schema.cpp
+/// file: conduit_schema.cpp
 ///
 //-----------------------------------------------------------------------------
-#include "Schema.hpp"
+#include "conduit_schema.hpp"
 
 //-----------------------------------------------------------------------------
 // -- standard lib includes -- 
@@ -57,9 +57,9 @@
 //-----------------------------------------------------------------------------
 // -- conduit includes -- 
 //-----------------------------------------------------------------------------
-#include "Generator.hpp"
-#include "Error.hpp"
-#include "Utils.hpp"
+#include "conduit_generator.hpp"
+#include "conduit_error.hpp"
+#include "conduit_utils.hpp"
 
 
 //-----------------------------------------------------------------------------
