+++ src/libs/conduit/conduit_schema.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Schema.hpp
+/// file: conduit_schema.hpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -62,9 +62,9 @@
 //-----------------------------------------------------------------------------
 // -- conduit includes -- 
 //-----------------------------------------------------------------------------
-#include "Core.hpp"
-#include "Endianness.hpp"
-#include "DataType.hpp"
+#include "conduit_core.hpp"
+#include "conduit_endianness.hpp"
+#include "conduit_data_type.hpp"
 
 
 //-----------------------------------------------------------------------------
