+++ src/libs/conduit/conduit_utils.cpp
@@ -44,11 +44,11 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Utils.cpp
+/// file: conduit_utils.cpp
 ///
 //-----------------------------------------------------------------------------
-#include "Utils.hpp"
-#include "Error.hpp"
+#include "conduit_utils.hpp"
+#include "conduit_error.hpp"
 
 //-----------------------------------------------------------------------------
 // -- standard lib includes --
