+++ src/libs/conduit/conduit_utils.hpp
@@ -44,7 +44,7 @@
 
 //-----------------------------------------------------------------------------
 ///
-/// file: Utils.hpp
+/// file: conduit_utils.hpp
 ///
 //-----------------------------------------------------------------------------
 
@@ -63,7 +63,7 @@
 //-----------------------------------------------------------------------------
 // -- conduit includes -- 
 //-----------------------------------------------------------------------------
-#include "Core.hpp"
+#include "conduit_core.hpp"
 
 
 //-----------------------------------------------------------------------------
