+++ src/libs/conduit/fortran/conduit_fortran_bitwidth_style_types.inc.in
@@ -44,7 +44,7 @@
 
 
 !------------------------------------------------------------------------------
-! file: Bitwidth_Style_Types.inc
+! file: conduit_fortran_bitwidth_style_types.inc
 !------------------------------------------------------------------------------
 
 !------------------------------------------------------------------------------
